UI Health, Progress, Thruster bars
Progress and Health bars function as I intended, thruster is almost ready.

Each bar has a UI Canvas in the prefab.

I createed two different test scenes. One is for the Slider health bar which I personally was not a fan of. The other is the FillerHealthTest which uses the Filled image type for the UI.

The buttons for each of the scene just are a way to show how the bars looks when they fill/empty.