Rock Debris (3 Sculpts, each with 2 sets of textures)
Here are some rocks to run into!  There's 3 different meshes and each one has 2 different colors, so hopefully that's enough variety.  I wanted to add in a little otter friend as a surprise, but getting the rocks in is more important at this point.  (that and my personal portfolio piece eeeek)